Fix non-native tests

+++ r1cs/gadgets/std/src/groups/nonnative/short_weierstrass/short_weierstrass_jacobian.rs
@@ -22,7 +22,7 @@ use crate::{
     },
     prelude::EqGadget,
     select::{CondSelectGadget, TwoBitLookupGadget},
-    uint8::UInt8,
+    UInt8,
     Assignment, ToBitsGadget, ToBytesGadget,
 };
 use std::{
